Rename vswhere.rs to vs_instances.rs
